Create DatePicker module to wrap react-datepicker
